Fix edition config for nightly
