+++ Cargo.toml
@@ -1,3 +1,4 @@
+cargo-features = ["edition2024"]
 [package]
 name = "safecmd"
 version = "0.1.0"
