+++ src/commands/cp/mod.rs
@@ -1,0 +1,3 @@
+pub fn run() {
+    println!("hello world");
+}
