+++ tests/trash_integration.rs
@@ -1,7 +1,6 @@
 use assert_cmd::prelude::*;
 use predicates;
 use std::fs::{self, File};
-use std::io::Write;
 use std::process::Command;
 use tempfile::tempdir;
